requirements.txt
# HTTP clients
aiohttp==3.9.1
httpx==0.26.0

# GitHub API wrapper
PyGithub==2.1.1

# OpenAI API client
openai==1.12.0

# Environment variable management
python-dotenv==1.0.0

# Async support
asyncio==3.4.3

# CLI interface
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
